ai-service/requirements.txt
# VerifyX AI Service - Python Dependencies
# ========================================

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# AI/ML Libraries
deepface>=0.0.79
opencv-python>=4.8.0
tf-keras>=2.16.0

# Image Processing
Pillow>=10.0.0

# Utilities
numpy>=1.26.0
python-dotenv>=1.0.0
requests>=2.31.0
